separation of file content and rework of mqtt
