+++ web/themes/dark19_01/processAllMqttMsg.js
@@ -1,226 +1,9 @@
 /**
- * Functions to update graph and values via MQTT
+ * Functions to update graph and gui values via MQTT-messages
  *
  * @author Kevin Wieland
  * @author Michael Ortenstein
  */
-var awattartime = [];
-var graphawattarprice;
-var doInterval;
-var initialread = 0;
-var graphloaded = 0;
-var boolDisplayHouseConsumption;
-var boolDisplayLoad1;
-var boolDisplayLp1Soc;
-var boolDisplayLoad2;
-var boolDisplayLp2Soc;
-var boolDisplayLp1;
-var boolDisplayLp2;
-var boolDisplayLp3;
-var boolDisplayLp4;
-var boolDisplayLp5;
-var boolDisplayLp6;
-var boolDisplayLp7;
-var boolDisplayLp8;
-var boolDisplayLpAll;
-var boolDisplaySpeicherSoc;
-var boolDisplaySpeicher;
-var boolDisplayEvu;
-var boolDisplayPv;
-var boolDisplayLegend;
-var boolDisplayLiveGraph;
-var datasend = 0;
-var all1 = 0;
-var all2 = 0;
-var all3 = 0;
-var all4 = 0;
-var all5 = 0;
-var all6 = 0;
-var all7 = 0;
-var all8 = 0;
-var all1p;
-var all2p;
-var all3p;
-var all4p;
-var all5p;
-var all6p;
-var all7p;
-var all8p;
-var hidehaus;
-var thevalues = [
-	"openWB/global/awattar/MaxPriceForCharging",
-	"openWB/global/awattar/pricelist",
-	"openWB/graph/lastlivevalues",
-	"openWB/graph/1alllivevalues",
-	"openWB/graph/2alllivevalues",
-	"openWB/graph/3alllivevalues",
-	"openWB/graph/4alllivevalues",
-	"openWB/graph/5alllivevalues",
-	"openWB/graph/6alllivevalues",
-	"openWB/graph/7alllivevalues",
-	"openWB/graph/8alllivevalues",
-	"openWB/graph/boolDisplayHouseConsumption",
-	"openWB/graph/boolDisplayLoad1",
-	"openWB/graph/boolDisplayLoad2",
-	"openWB/graph/boolDisplayLp1Soc",
-	"openWB/graph/boolDisplayLp2Soc",
-	"openWB/graph/boolDisplayLp1",
-	"openWB/graph/boolDisplayLp2",
-	"openWB/graph/boolDisplayLp3",
-	"openWB/graph/boolDisplayLp4",
-	"openWB/graph/boolDisplayLp5",
-	"openWB/graph/boolDisplayLp6",
-	"openWB/graph/boolDisplayLp7",
-	"openWB/graph/boolDisplayLp8",
-	"openWB/graph/boolDisplayLpAll",
-	"openWB/graph/boolDisplaySpeicherSoc",
-	"openWB/graph/boolDisplaySpeicher",
-	"openWB/graph/boolDisplayEvu",
-	"openWB/graph/boolDisplayLegend",
-	"openWB/graph/boolDisplayLiveGraph",
-	"openWB/graph/boolDisplayPv",
-	"openWB/evu/W",
-	"openWB/global/WHouseConsumption",
-	"openWB/pv/W",
-	"openWB/pv/DailyYieldKwh",
-	"openWB/lp/1/%Soc",
-	"openWB/lp/2/%Soc",
-	// heute geladene kWh ... nicht benutzt im Theme
-	"openWB/lp/1/kWhDailyCharged",
-	"openWB/lp/2/kWhDailyCharged",
-	"openWB/lp/3/kWhDailyCharged",
-	// geladene kWh des aktuellen Ladesegments
-	"openWB/lp/1/kWhActualCharged",
-	"openWB/lp/2/kWhActualCharged",
-	"openWB/lp/3/kWhActualCharged",
-	// geladene kWh seit anstecken des EV
-	"openWB/lp/1/kWhChargedSincePlugged",
-	"openWB/lp/2/kWhChargedSincePlugged",
-	"openWB/lp/3/kWhChargedSincePlugged",
-	"openWB/lp/4/kWhChargedSincePlugged",
-	"openWB/lp/5/kWhChargedSincePlugged",
-	"openWB/lp/6/kWhChargedSincePlugged",
-	"openWB/lp/7/kWhChargedSincePlugged",
-	"openWB/lp/8/kWhChargedSincePlugged",
-	// Ladeleistung am LP
-	"openWB/lp/1/W",
-	"openWB/lp/2/W",
-	"openWB/lp/3/W",
-	"openWB/lp/4/W",
-	"openWB/lp/5/W",
-	"openWB/lp/6/W",
-	"openWB/lp/7/W",
-	"openWB/lp/8/W",
-	"openWB/lp/1/boolPlugStat",
-	"openWB/lp/2/boolPlugStat",
-	"openWB/lp/3/boolPlugStat",
-	"openWB/lp/4/boolPlugStat",
-	"openWB/lp/5/boolPlugStat",
-	"openWB/lp/6/boolPlugStat",
-	"openWB/lp/7/boolPlugStat",
-	"openWB/lp/8/boolPlugStat",
-	"openWB/lp/1/boolChargeStat",
-	"openWB/lp/2/boolChargeStat",
-	"openWB/lp/3/boolChargeStat",
-	"openWB/lp/4/boolChargeStat",
-	"openWB/lp/5/boolChargeStat",
-	"openWB/lp/6/boolChargeStat",
-	"openWB/lp/7/boolChargeStat",
-	"openWB/lp/8/boolChargeStat",
-	"openWB/lp/1/boolSocConfigured",
-	"openWB/lp/2/boolSocConfigured",
-	"openWB/lp/1/AConfigured",
-	"openWB/lp/2/AConfigured",
-	"openWB/lp/3/AConfigured",
-	"openWB/lp/8/AConfigured",
-	"openWB/lp/4/AConfigured",
-	"openWB/lp/5/AConfigured",
-	"openWB/lp/6/AConfigured",
-	"openWB/lp/7/AConfigured",
-	"openWB/lp/1/TimeRemaining",
-	"openWB/lp/2/TimeRemaining",
-	"openWB/lp/3/TimeRemaining",
-	"openWB/lp/1/kmCharged",
-	"openWB/lp/2/kmCharged",
-	"openWB/lp/3/kmCharged",
-	"openWB/lp/1/ChargeStatus",
-	"openWB/lp/2/ChargeStatus",
-	"openWB/lp/3/ChargeStatus",
-	"openWB/lp/4/ChargeStatus",
-	"openWB/lp/5/ChargeStatus",
-	"openWB/lp/6/ChargeStatus",
-	"openWB/lp/7/ChargeStatus",
-	"openWB/lp/8/ChargeStatus",
-	"openWB/global/ChargeMode",
-	"openWB/global/WAllChargePoints",
-	"openWB/housebattery/boolHouseBatteryConfigured",
-	"openWB/housebattery/W",
-	"openWB/housebattery/%Soc",
-	"openWB/global/strLastmanagementActive",
-	"openWB/lp/1/boolChargePointConfigured",
-	"openWB/lp/2/boolChargePointConfigured",
-	"openWB/lp/3/boolChargePointConfigured",
-	"openWB/lp/4/boolChargePointConfigured",
-	"openWB/lp/5/boolChargePointConfigured",
-	"openWB/lp/6/boolChargePointConfigured",
-	"openWB/lp/7/boolChargePointConfigured",
-	"openWB/lp/8/boolChargePointConfigured",
-	"openWB/lp/1/boolDirectChargeMode_none_kwh_soc",
-	"openWB/lp/2/boolDirectChargeMode_none_kwh_soc",
-	"openWB/lp/3/boolDirectChargeMode_none_kwh_soc",
-	"openWB/lp/4/boolDirectChargeMode_none_kwh_soc",
-	"openWB/lp/5/boolDirectChargeMode_none_kwh_soc",
-	"openWB/lp/6/boolDirectChargeMode_none_kwh_soc",
-	"openWB/lp/7/boolDirectChargeMode_none_kwh_soc",
-	"openWB/lp/8/boolDirectChargeMode_none_kwh_soc",
-
-	"openWB/lp/1/boolChargeAtNight",
-	"openWB/lp/2/boolChargeAtNight",
-
-	"openWB/lp/1/ChargePointEnabled",
-	"openWB/lp/2/ChargePointEnabled",
-	"openWB/lp/3/ChargePointEnabled",
-	"openWB/lp/4/ChargePointEnabled",
-	"openWB/lp/5/ChargePointEnabled",
-	"openWB/lp/6/ChargePointEnabled",
-	"openWB/lp/7/ChargePointEnabled",
-	"openWB/lp/8/ChargePointEnabled",
-	"openWB/lp/1/strChargePointName",
-	"openWB/lp/2/strChargePointName",
-	"openWB/lp/3/strChargePointName",
-	"openWB/lp/4/strChargePointName",
-	"openWB/lp/5/strChargePointName",
-	"openWB/lp/6/strChargePointName",
-	"openWB/lp/7/strChargePointName",
-	"openWB/lp/8/strChargePointName",
-	"openWB/lp/1/AutolockConfigured",
-	"openWB/lp/2/AutolockConfigured",
-	"openWB/lp/3/AutolockConfigured",
-	"openWB/lp/4/AutolockConfigured",
-	"openWB/lp/5/AutolockConfigured",
-	"openWB/lp/6/AutolockConfigured",
-	"openWB/lp/7/AutolockConfigured",
-	"openWB/lp/8/AutolockConfigured",
-	"openWB/lp/1/AutolockStatus",
-	"openWB/lp/2/AutolockStatus",
-	"openWB/lp/3/AutolockStatus",
-	"openWB/lp/4/AutolockStatus",
-	"openWB/lp/5/AutolockStatus",
-	"openWB/lp/6/AutolockStatus",
-	"openWB/lp/7/AutolockStatus",
-	"openWB/lp/8/AutolockStatus",
-	"openWB/lp/1/ADirectModeAmps",
-	"openWB/lp/2/ADirectModeAmps",
-	"openWB/lp/3/ADirectModeAmps",
-	"openWB/lp/4/ADirectModeAmps",
-	"openWB/lp/5/ADirectModeAmps",
-	"openWB/lp/6/ADirectModeAmps",
-	"openWB/lp/7/ADirectModeAmps",
-	"openWB/lp/8/ADirectModeAmps",
-	"openWB/system/Timestamp"
-];
-var clientuid = Math.random().toString(36).replace(/[^a-z]+/g, "").substr(0, 5);
-var client = new Messaging.Client(location.host, 9001, clientuid);
 
 function getCol(matrix, col){
 	var column = [];
@@ -1036,184 +819,3 @@ function processLpMessages(mqttmsg, mqttpayload) {
 		}
 	}
 }
-
-//Gets  called if the websocket/mqtt connection gets disconnected for any reason
-client.onConnectionLost = function (responseObject) {
-	client.connect(options);
-};
-//Gets called whenever you receive a message
-client.onMessageArrived = function (message) {
-	handlevar(message.destinationName, message.payloadString);
-};
-var retries = 0;
-
-//Connect Options
-var options = {
-	timeout: 5,
-	//Gets Called if the connection has sucessfully been established
-	onSuccess: function () {
-		retries = 0;
-		thevalues.forEach(function(thevar) {
-			client.subscribe(thevar, {qos: 0});
-		});
-	},
-	//Gets Called if the connection could not be established
-	onFailure: function (message) {
-		client.connect(options);
-	}
-};
-
-//Creates a new Messaging.Message Object and sends it
-var publish = function (payload, topic) {
-	var message = new Messaging.Message(payload);
-	message.destinationName = topic;
-	message.qos = 2;
-	message.retained = true;
-	client.send(message);
-}
-
-client.connect(options);
-
-function graphoptionclick() {
-	if ( document.getElementById("graphoptiondiv").style.display === "none") {
-		document.getElementById("graphoptiondiv").style.display = "block";
-	} else {
-		document.getElementById("graphoptiondiv").style.display = "none";
-	}
-}
-
-function AwattarMaxPriceClick() {
-	publish(document.getElementById("awattar1l").innerHTML,"openWB/set/awattar/MaxPriceForCharging");
-}
-function lp1DirectChargeAmpsClick() {
-	publish(document.getElementById("sofortlllp1l").innerHTML,"openWB/set/lp1/DirectChargeAmps");
-}
-
-function lp2DirectChargeAmpsClick() {
-	publish(document.getElementById("sofortlllp2l").innerHTML,"openWB/set/lp2/DirectChargeAmps");
-}
-
-function lp3DirectChargeAmpsClick() {
-	publish(document.getElementById("sofortlllp3l").innerHTML,"openWB/set/lp3/DirectChargeAmps");
-}
-
-function lp4DirectChargeAmpsClick() {
-	publish(document.getElementById("sofortlllp4l").innerHTML,"openWB/set/lp4/DirectChargeAmps");
-}
-
-function lp5DirectChargeAmpsClick() {
-	publish(document.getElementById("sofortlllp5l").innerHTML,"openWB/set/lp5/DirectChargeAmps");
-}
-
-function lp6DirectChargeAmpsClick() {
-	publish(document.getElementById("sofortlllp6l").innerHTML,"openWB/set/lp6/DirectChargeAmps");
-}
-
-function lp7DirectChargeAmpsClick() {
-	publish(document.getElementById("sofortlllp7l").innerHTML,"openWB/set/lp7/DirectChargeAmps");
-}
-
-function lp8DirectChargeAmpsClick() {
-	publish(document.getElementById("sofortlllp8l").innerHTML,"openWB/set/lp8/DirectChargeAmps");
-}
-
-
-function renewMQTTclick() {
-	publish("1","openWB/set/RenewMQTT");
-	alert("Erneuern der Werte initiert, dies dauert ca 15-20 Sekunden.");
-}
-
-function subscribeMqttGraphSegments() {
-	for (var segments = 1; segments < 9; segments++) {
-		client.subscribe("openWB/graph/" + segments + "alllivevalues", {qos: 0});
-	}
-}
-
-function unsubscribeMqttGraphSegments() {
-	for (var segments = 1; segments < 9; segments++) {
-		client.unsubscribe("openWB/graph/" + segments + "alllivevalues");
-	}
-}
-
-function putgraphtogether() {
-	if ( (all1 == 1) && (all2 == 1) && (all3 == 1) && (all4 == 1) && (all5 == 1) && (all6 == 1) && (all7 == 1) && (all8 == 1) ){
-		var alldata = all1p + "\n" + all2p + "\n" + all3p + "\n" + all4p + "\n" + all5p + "\n" + all6p + "\n" + all7p + "\n" + all8p;
-		alldata = alldata.replace(/^\s*[\n]/gm, "");
-		alldata = alldata.replace(/^\s*-[\n]/gm, "");
-		var csvData = [];
-		var rawcsv = alldata.split(/\r?\n|\r/);
-		for (var i = 0; i < rawcsv.length; i++) {
-			  csvData.push(rawcsv[i].split(","));
-		}
-		csvData.pop();
-		// Retrived data from csv file content
-		var splittime = [];
-		getCol(csvData, 0).forEach(function(zeit){
-			splittime.push(zeit.substring(0, zeit.length -3));
-		});
-		atime = splittime;
-		//atime = getCol(csvData, 0);
-		abezug = convertToKw(getCol(csvData, 1));
-		alpa = convertToKw(getCol(csvData, 2));
-		apv = convertToKw(getCol(csvData, 3));
-		alp1 = convertToKw(getCol(csvData, 4));
-		alp2 = convertToKw(getCol(csvData, 5));
-		aspeicherl = convertToKw(getCol(csvData, 7));
-		aspeichersoc = getCol(csvData, 8);
-		asoc = getCol(csvData, 9);
-		asoc1 = getCol(csvData, 10);
-		ahausverbrauch = convertToKw(getCol(csvData, 11));
-		averbraucher1 = convertToKw(getCol(csvData, 12));
-		averbraucher2 = convertToKw(getCol(csvData, 13));
-		alp3 = convertToKw(getCol(csvData, 14));
-		alp4 = convertToKw(getCol(csvData, 15));
-		alp5 = convertToKw(getCol(csvData, 16));
-		alp6 = convertToKw(getCol(csvData, 17));
-		alp7 = convertToKw(getCol(csvData, 18));
-		alp8 = convertToKw(getCol(csvData, 19));
-		initialread = 1 ;
-
-		// after receipt of all 8 first data segments, unsubscribe from these topics to save bandwidth
-		unsubscribeMqttGraphSegments();
-
-		checkgraphload();
-	}
-}  // end putgraphtogether
-
-
-$(window).focus(function() {
-    // if the browser window gets focus again after being blurred,
-    // check if mqtt segments for graph need to be subsribed again
-    if ( initialread == 0 ) {
-		subscribeMqttGraphSegments();
-	}
-});
-
-function getHookStatus(dataURL) {
-	// read dataURL filecontent and return it
-	return $.get(dataURL);
-}
-
-function displayHookStatus(hookNumber) {
-	var element = "#hook"+hookNumber+"div";
-	getHookStatus("/openWB/ramdisk/hook"+hookNumber+"akt").done(function(result) {
-		if ( result == 1 ) {
-			$(element).removeClass("bg-danger");
-			$(element).addClass("bg-success");
-		} else {
-			$(element).removeClass("bg-success");
-			$(element).addClass("bg-danger");
-		}
-	});
-}
-
-function processAllHooks() {
-	for (numberOfHook=1; numberOfHook<=3; numberOfHook++) {
-		displayHookStatus(numberOfHook);
-	}
-}
-
-$(document).ready(function(){
-	doInterval = setInterval(processAllHooks, 5000);
-	processAllHooks();
-});
